Simplify calendar adapter creation
