+++ web/src/components/__tests__/PeriodNavigation.test.tsx
@@ -88,9 +88,7 @@ vi.mock("@/utils/calendar", () => ({
   isLocalDateString: (value: unknown) =>
     typeof value === "string" && /^\d{4}-\d{2}-\d{2}$/.test(value),
   parseLocalDateString: (value: string) => new Date(`${value}T00:00:00`),
-  CalendarAdapterFactory: {
-    create: vi.fn(() => calendarAdapterMock),
-  },
+  createCalendarAdapter: vi.fn(() => calendarAdapterMock),
 }));
 
 vi.mock("@/hooks/queries/usePreferenceWithBootstrap", () => ({
