+++ web/src/pages/__tests__/PlanningPage.integration.test.tsx
@@ -93,9 +93,7 @@ vi.mock("@/utils/calendar", () => ({
   isLocalDateString: (value: unknown) =>
     typeof value === "string" && /^\d{4}-\d{2}-\d{2}$/.test(value),
   parseLocalDateString: (value: string) => new Date(`${value}T00:00:00`),
-  CalendarAdapterFactory: {
-    create: vi.fn(() => calendarAdapter),
-  },
+  createCalendarAdapter: vi.fn(() => calendarAdapter),
 }));
 
 interface UsePlanningTasksMockResult {
